% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_bootstrap.R
\name{para_bootstrap}
\alias{para_bootstrap}
\title{Title function of Parametric bootstrap}
\usage{
para_bootstrap(B, Dataset, para)
}
\arguments{
\item{B}{iteration of bootstrap}

\item{Dataset}{the dataset of my own example - "gopher_tortoise" dataset}

\item{para}{the value of the coefficient of the predictor -- "prev". For example, tortoise_fit$beta\link{2}, or 0}
}
\value{
return the bootstrap which contains B, effect, group. term, estimates, standard error, statistic, and p-value.
}
\description{
Title function of Parametric bootstrap
}
\examples{
## choosing different B into the para_bootstrap function
data(tortoise)
tortoise_fit = run_model(tortoise, "tortoise")
para_bootstrap(100, tortoise, tortoise_fit$beta[2])
}
